Update patch set 6
Patch Set 6:

> Was there a specific reason to update the dev version?

The previous version was 2.3.4. So I needed to set a dev version to compile it.  Meanwhile another patch released which changed it to 2.3.4-dev. But if I don't know it wrong,  the dev version should include the target relase which is 2.3.5-dev, not 2.3.4-dev. That is why i updated it to 2.3.5-dev. If I'm wrong, I can change it to accept the incoming change.

Patch-set: 6
Reviewer: Gerrit User 1000218 <1000218@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0