Update patch set 2
Patch Set 2:

> (1 comment)

makes sense to augment the proto with instance_Id.   How the adapter instance Id gets set?   On a container restart, does it reuse the same ID?  When an adapter goes down and re-register, the Core will send the reconcile device request on the adapter topic and all of them will receive it.   Which devices will the new adapter reconcile and how about the ones that did not restart?

Patch-set: 2