Update patch set 2
Patch Set 2:

> > > (1 comment)
 > >
 > > makes sense to augment the proto with instance_Id.   How the
 > > adapter instance Id gets set?   On a container restart, does it
 > > reuse the same ID?  When an adapter goes down and re-register,
 > the
 > > Core will send the reconcile device request on the adapter topic
 > > and all of them will receive it.   Which devices will the new
 > > adapter reconcile and how about the ones that did not restart?
 > 
 > well, it depends on the implementation in the adapter. I just
 > needed to slightly modify the usage of instance_id parameter in
 > open-onu-adapter. It basically get the OS.name as an environment
 > parameter and in k8s deployment it's the name of the node and it
 > doesn't change between restarts.
 > As you said when a container is restarted, all of them may get the
 > reconcile request. It might be a good idea to insert the
 > instance_id into the reconcile request and only the others may
 > ignore the request. However, in my setup this is not a problem
 > since I've been trying kafka's HashPartitioning algorithm and set
 > the key=device_id.

It's a good idea to use k8s to maintain the instance_id across restarts.

Within the kafka library, all message sent to kafka uses the device-id as a key for partionning.  While it's there the Core uses group consumer meaning every Core listening on that topic would get the message.  That works well when we have 2-core in an HA pair.   If instead of group consumer you use partition consumer then only one adapter will pull a message from kafka at a time.  So your approach will work.

The only difficulty is on restart.  When an adapter restarts how does it know which partition/key it needs to listen to?  And if there are multiple adapters that restarted at the same time?

The Core does not keep the association between a device and the adapter instance.  I don't think the Core needs to do that.   This belong to the realm of the adapters.   Instead of using deviceId as the key, would it be easy to use the adapter instanceId as the key?

Patch-set: 2