[VOL-2790] Fix delete pipeline issue
During a delete several handlers get executed in a pipeline.  If
a handler returns an error then the remaining pipeline will be
ignored.   This change addresses the issue during a deletion
flow where a handler is currently returning an error when there
is not really a failure, i.e. the device upon which the handler is
executing may already have been deleted.

In the future, we need to implement specific error objects and
decide whether an error need to be thrown or not.   Won't make a
difference for the delete scenario as the error seen is that the
device is not found.

Change-Id: I3e0755f78d53acdf3b7600c41aef0b5f65c4af02