Update patch set 4
Patch Set 4:

The change to use a generated logical device Id and keep the mac address as the DPID looks good.  We previously was using generated IDs for both of them. Two years ago we made a change to use the MAC address as the DPID.   While there were no clear requirements to stop generating the logical device ID, out of convenience both IDs were set to be the same.

Patch-set: 4
Label: Code-Review=0