Update patch set 2
Patch Set 2:

> Btw, when running this patch in my local environment panics with
 > concurrent map read/write.  I guess this is because it is still in
 > progress. Some adjustment of code surrounding this change would be
 > needed.

Actually i haven't tested it enough yet. I uploaded it at once to let you for a pre-review. So,  if this will not be an accepted way of doing things , then I wouldn't spend to much time to test it. That is why it is still in WIP status. But I will be happy if you guys also test it. This needs to be tested well before merged.

Patch-set: 2