Update patch set 16
Patch Set 16:

(8 comments)

I'm not yet convinced that concurrency rwt mutexes is guaranteed to be correct in all cases.  The locks seem to be used inconsistently, and seem to attempt to guarantee different things in different places.

The locks surrounding this are much more complicated than they were before.  We need to guarantee both in-memory consistency & write-to-db consistency (no parallel writes to the same key).  I'm just not sure we're at that point yet.

Not sure we should gate the patch on this, but it _must_ be addressed at some point.

Patch-set: 16
Label: Code-Review=0