Update patch set 1
Patch Set 1:

This is difficult to review as-is.  
Some functions are moved around, and they so cannot be diffed with previous versions.

It would be better to break down agent.go into multiple files in one patch, and make actual functionality changes in a separate patch.

Is this feasible?

Patch-set: 1
Label: Verified=0