Update patch set 7
Patch Set 7:

> > Does voltha-lib-go actually get the KV data prefix from the
 > > environment variable?   If that's the case that's a bug since
 > it's
 > > a library and should just accept whatever prefix is passed to it.
 > 
 > That was a suggestion from David and it did make sense since it's a
 > library, each component uses it should not have different
 > command-line parameters to configure the kv-path. It does no effect
 > right now because every component has it's own docker instance. If
 > one day we decide to run components in one docker instance then
 > having a common kv store prefix as an environment variable makes
 > sense to me. I can't say it's a bug. But if someone wants to use a
 > command-line parameter to configure it, it should take the
 > precedence over the env variable.

I see the benefit but removing the KV prefix attributes from the config file will not allow someone to configure it via a command-line parameter, other than a env variable.   If its the library code that is reading the env variable then I have an issue with that.  I would prefer that when a component uses the library it passes in the env variable.

Patch-set: 7