[VOL-3245] Consume port capability on port creation
This commit consists of the changes needed in the Core to build
a logical port without having to request port capability from
the adapters.

This commit requires changes in the OLT and ONU adapters to be
completed.  These are tracked under jira
https://jira.opencord.org/browse/VOL-3202. Until the adapters
changes are committed this commit will not pass jenkins.

Change-Id: I0cae167375b9a8c67a83202e25abcec6ae013a88