Update patch set 6
Patch Set 6:

Despite this patch has been merged, I want to make some points clearer for Khenaidoo.

I understand your concern and these are mines:

GroupDelete is the only method where a group and related flows are handled together. Otherwise, multicast flows and groups are handled in separate workflows. That is why corresponding code block is not repeated in other functions, it is specific to GroupDelete function.

Groups are handled by groupAdd, groupModify, groupDelete, and flows are handled by flowAdd, flowDelete and flowDeleteStrict methods of logical_agent.

Let's suppose we move the following code block to processMulticastFlow() method of flow_decomposer. 

	//add groups to deviceRules
	for _, groupEntry := range affectedGroups {
		fg := fu.NewFlowsAndGroups()
		fg.AddGroup(groupEntry)
		deviceRules.AddFlowsAndGroup(agent.rootDeviceID, fg)
	}
	
In this case, the resulting behavior will change depending on the context. In flowAdd workflow, this change will result in a group creation request, in addition the flow creation request, to the adapter for this group referenced by the flow. However, this is not the way groups should be created. They arrives at the VOLTHA Core via GroupMod messages and handled by the groupAdd workflow.

On flowDelete workflow, this change in the flow_decomposer will result in sending a group delete request to the adapter along with the flow delete request. However, this is not the way groups should be deleted. A GroupMod message is expected to trigger a group delete request to the adapter. 

These are the concerns that hinder me from encapsulating this behavior in the flow_decomposer. I hope it makes sense.

Thanks,

Patch-set: 6