Update patch set 2
Patch Set 2:

@Khen

The spec says: "When a flow entry is removed, either by the controller or the flow expiry mechanism, the switch must
check the flow entry’s OFPFF_SEND_FLOW_REM flag. If this flag is set, the switch must send a flow removed
message to the controller."

And the ofAgent is already correctly sending the response, but what happens is that in some case the response is sent before the flow is actually removed. To elaborate, in the ATT workflow:
- an EAPOL flow with the default meter is created (let's say the meterId is 1)
- when we provision the subscriber the EAPOL flow is removed a new EAPOL flow, this time with a different meterId is pushed
- If the second EAPOL flow is received before the first one is removed we have a meter mismatch in the device

What we are trying to ensure here is that if the controller care about the flow removal response we should guarantee is sent only after the flow is actually removed.

Patch-set: 2