Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Khen
> 
> The spec says: "When a flow entry is removed, either by the controller or the flow expiry mechanism, the switch must
> check the flow entry’s OFPFF_SEND_FLOW_REM flag. If this flag is set, the switch must send a flow removed
> message to the controller."
> 
> And the ofAgent is already correctly sending the response, but what happens is that in some case the response is sent before the flow is actually removed. To elaborate, in the ATT workflow:
> - an EAPOL flow with the default meter is created (let's say the meterId is 1)
> - when we provision the subscriber the EAPOL flow is removed a new EAPOL flow, this time with a different meterId is pushed
> - If the second EAPOL flow is received before the first one is removed we have a meter mismatch in the device
> 
> What we are trying to ensure here is that if the controller care about the flow removal response we should guarantee is sent only after the flow is actually removed.

@Matteo

I understand the reasoning behind this.  My question is should that flow-removed response be synchronous (as being suggested) or should it be asynchronous.  Should the provisioning mechanism wait for a response (when it sets the OFPFF_SEND_FLOW_REM flag) before sending the new flow?    

As i mentioned, i am not against this implementation.   I am just trying to see if there is a better way to handle this situation including future ones.  For instance, if there is a need to add a flow after another flow is added would we need another flag for the Core to manage this or should we instead have that logic controlled by the component that needs to have this sequencing of flows?

Patch-set: 2