Update patch set 5
Patch Set 5:

The unit test fails due to a panic in etcd library. I do not see this issue on my mac and the change does not seem to be related to my change at all. I am going to "recheck" this patch to see if I get a +1.

Patch-set: 5
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>