Update patch set 8
Patch Set 8:

As per the error handling document (https://docs.google.com/document/d/1sOv5R7iOBTlzXcCVaUfOOzd5617ISQGnseadhx8pb3w/edit?pli=1#), when a device is in RECONCILING state, no operations on that device is allowed, other than the delete operation and operational/connection state changes from the adapter as part of the reconcile process.

In this commit there is a check only when a device is being "enabled" (applies only if the device was disabled before, otherwise the adapter would have no knowledge of that device.   Would be useful to have a common check in all methods where the admin state of the device is being changed.

Patch-set: 8
Reviewer: Gerrit User 1000056 <1000056@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>