Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> There is the agent_device_update.go file where any update to a device gets published to logs.  This seems the ideal place to check for any state change (code already present) and if there is to publish to kafka as well.

The logDeviceUpdate function in  agent_device_update.go is only logging the admin state and the function is called in more than 50 places only for logging(whit nil states), where no updates made on device. So I don't think it is a good place to do this.

Patch-set: 8