Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Hi Esin,
> 
> I have a patch that takes care of a core restart (along with some other changes) which is being tested under load condition.  There are some changes required in your patch that I have in mine.
> 
> If you want I can send you my patch to integrate with yours or I can submit mine.  I am ok either way.
> 
> Khen

Hi Ken,

If you have a patch which have already taken care of the same issue, I really appreciate it. I think it would be good to have your patch first. Could you add me and Serkant as reviewer so that we can take a look at it. We can consider my patch if we still need it after your updates.

Thanks,
Esin

Patch-set: 2