Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> my understanding is that we woudl never retry, this does retry untile it gets a RECONCILING_FAILED, also independently of this patch, I'm still seeing the retry arrive each second, is there an issue with the timer ?

I thought the plan was to retry until there is a failure set by the adapter this way if the issue was a comms issue then when the comms issues was addressed the reconcile could proceed. The issue was that once the adapter sent a RECONCILING_FAILED as the oper state then there was no point to retry.

As for the seeing the a retry every second, not sure and that may be a separate issue.

Patch-set: 4