VOL-4511: Logical Meters are not deleted from the kv store after olt device delete some times
After olt device delete, the logical meters are cleaned up, but before
the logical device could be deleted, ONOS re-installs the meters and
this becomes a stale entry on the KV store.
The fix here is to delay logical meter deletion until logical device
is deleted from the KV store.

Change-Id: I4010da6dead26b481e56ac29d576d8f3e48a936a